+++ src/events/index.ts
@@ -1,0 +1,2 @@
+export type { ManagedSerialPortsEvent } from "@/events/managed-serial-ports";
+export type { PacketEvent } from "@/events/packet";
